Fix the player being locked when Strength pushes a blocked boulder

+++ include/qol_field_moves.h
@@ -14,7 +14,7 @@ void RemoveRelevantSurfFieldEffect(void);
 void Task_SurfToolFieldEffect(u8 taskId);
 
 u32 CanUseStrength(u8);
-u32 UseStrength(u32, u8, u8, u8);
+u32 UseStrength(u32, s16, s16, u8);
 
 void FldEff_UseFlashTool(void);
 u32 CanUseFlash(void);
